+++ app/(route)/my/reservations/page.tsx
@@ -1,18 +1,10 @@
 import getReservations from '@/_apis/reservations/getReservations';
 import { dehydrate, HydrationBoundary, QueryClient } from '@tanstack/react-query';
 
-import type { ReservationStatus } from '@/_types/myReservations';
-
 import CommonLayout from '@/_components/CommonLayout';
 import StickyLayout from '@/_components/SideStickyLayout';
 
 import ReservationClient from './_components/ReservationClient';
-
-export interface ReservationParams {
-  cursorId?: number;
-  size?: number;
-  status?: ReservationStatus;
-}
 
 async function MyReservations() {
   const queryClient = new QueryClient();
